requirements.txt
asgiref==3.2.3
atomicwrites==1.4.0
attrs==19.3.0
bson==0.5.8
certifi==2019.11.28
colorama==0.4.3
dataclasses==0.6
Django==3.0.6
django-appconf==1.0.4
django-ckeditor==5.9.0
django-filter==2.2.0
django-js-asset==1.2.2
django-suit==0.2.28
djangorestframework==3.11.0
Markdown==3.1.1
more-itertools==8.3.0
packaging==20.4
Pillow==7.1.2
pipenv==2018.11.26
pluggy==0.13.1
psycopg2==2.8.5
py==1.8.1
pycodestyle==2.5.0
pyparsing==2.4.7
pytest==5.4.2
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.2.4
virtualenv==16.7.9
virtualenv-clone==0.5.3
wcwidth==0.1.9
